Conflicts:
	.idea/artifacts/ij_webapp_server_war.xml
	.idea/artifacts/ij_webapp_server_war_exploded.xml
	.idea/artifacts/ij_webapp_system_war.xml
	.idea/artifacts/ij_webapp_system_war_exploded.xml
	.idea/compiler.xml
	.idea/encodings.xml
	ij.system.impl/pom.xml
	ij.user.api/src/main/java/net/moonlithome/game/server/account/mapper/AccountOperationMapper.xml
	ij.user.api/src/main/java/net/moonlithome/game/server/account/service/AccountOperationService.java
	ij.user.api/src/main/java/net/moonlithome/game/server/user/service/UserOperationService.java
	ij.user.impl/src/main/java/net/moonlithome/game/server/account/serviceImpl/AccountOperationServiceImpl.java
	ij.user.impl/src/main/java/net/moonlithome/game/server/user/serviceImpl/UserServiceImpl.java
	ij.webapp.server/pom.xml
	ij.webapp.server/src/main/java/net/moonlithome/game/server/account/action/AccountOperationAction.java
	ij.webapp.server/src/main/java/net/moonlithome/game/server/user/action/UserOperationAction.java
	ij.webapp.server/src/main/resources/applicationContext.xml
	ij.webapp.server/target/classes/applicationContext.xml
	ij.webapp.system/src/main/resources/applicationContext.xml

+++ ij.webapp.system/src/main/java/net/moonlithome/game/system/action/UserLevelManageAction.java
@@ -16,6 +16,7 @@ import javax.servlet.http.HttpServletResponse;
  * Created by MF on 3/15/2015.
  */
 @Controller
+@RequestMapping("/system")
 public class UserLevelManageAction extends BaseAction{
 
     @Autowired
@@ -24,34 +25,37 @@ public class UserLevelManageAction extends BaseAction{
     @RequestMapping("/addUserLevel")
     public void addUserLevel(HttpServletRequest request, HttpServletResponse response) throws Exception {
         UserLevelDto userLevelDto = JsonBeanUtil.getBeanFromJson(getJsonFromRequest(request), UserLevelDto.class);
-        //TODO check session
-        if(userLevelDto != null){
-            userLevelManageDao.addUserLevel(userLevelDto);
+        if(permission_check()) {
+            if (userLevelDto != null) {
+                userLevelManageDao.addUserLevel(userLevelDto);
+            }
+            BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
+            responseJson(response, baseCommunicationDto);
         }
-        BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
-        responseJson(response, baseCommunicationDto);
     }
 
     @RequestMapping("/updateUserLevel")
     public void updateUserLevel(HttpServletRequest request, HttpServletResponse response) throws Exception {
         UserLevelDto userLevelDto = JsonBeanUtil.getBeanFromJson(getJsonFromRequest(request), UserLevelDto.class);
-        //TODO check session
-        if(userLevelDto != null){
-            userLevelManageDao.updateUserLevel(userLevelDto);
+        if(permission_check()) {
+            if (userLevelDto != null) {
+                userLevelManageDao.updateUserLevel(userLevelDto);
+            }
+            BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
+            responseJson(response, baseCommunicationDto);
         }
-        BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
-        responseJson(response, baseCommunicationDto);
     }
 
     @RequestMapping("/removeUserLevel")
     public void removeUserLevel(HttpServletRequest request, HttpServletResponse response) throws Exception {
         UserLevelDto userLevelDto = JsonBeanUtil.getBeanFromJson(getJsonFromRequest(request), UserLevelDto.class);
-        //TODO check session
-        if(userLevelDto != null){
-            userLevelManageDao.removeUserLevel(userLevelDto.getId());
+        if(permission_check()) {
+            if (userLevelDto != null) {
+                userLevelManageDao.removeUserLevel(userLevelDto.getId());
+            }
+            BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
+            responseJson(response, baseCommunicationDto);
         }
-        BaseCommunicationDto baseCommunicationDto = new BaseCommunicationDto();
-        responseJson(response, baseCommunicationDto);
     }
 
     @RequestMapping("/getUserLevel")
